move into modules and run
