+++ crates/node/src/metrics.rs
@@ -1,7 +1,9 @@
-use std::sync::LazyLock;
+use std::{sync::LazyLock, time::Duration};
 
-pub(crate) mod task_metrics;
 pub(crate) mod tokio_runtime_metrics;
+pub(crate) mod tokio_task_metrics;
+
+const MONITOR_SAMPLE_DURATION: Duration = Duration::from_secs(10);
 
 pub static MPC_NUM_TRIPLES_GENERATED: LazyLock<prometheus::IntCounter> = LazyLock::new(|| {
     prometheus::register_int_counter!(
