+++ crates/node/src/providers/robust_ecdsa.rs
@@ -7,7 +7,7 @@ use std::collections::HashMap;
 
 use crate::config::{ConfigFile, MpcConfig, ParticipantsConfig};
 use crate::db::SecretDB;
-use crate::metrics::task_metrics::ROBUST_ECDSA_TASK_MONITORS;
+use crate::metrics::tokio_task_metrics::ROBUST_ECDSA_TASK_MONITORS;
 use crate::network::{MeshNetworkClient, NetworkTaskChannel};
 use crate::primitives::{MpcTaskId, UniqueId};
 use crate::providers::{EcdsaSignatureProvider, SignatureProvider};
